requirements.txt
#python==3.6.9
h5py==2.10.0
matplotlib==3.2.0
notebook==6.0.3
numpy==1.18.4
opencv-python==4.0.0.21
pandas==1.0.4
pickle==4.0
Pillow==7.0.0
psutil==5.6.7
pymc3==3.7
scikit-image==0.14.2
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.9.0
sklearn==0.0
statsmodels==0.11.1
tensorboard==1.8.0
tensorboardX==2.0
#torch==1.4.0
torchfile==0.1.0
#torchvision==0.4.2
tqdm==4.31.1
